arm64: insn: fix ADRP offset manipulation
ADRP instructions encode a scale 21-bit isigned immediate which can
represent offsets in the interval [-SZ_4G, SZ_4G - SZ_4K].

The aarch64_insn_adrp_get_offset() and aarch64_insn_adrp_set_offset()
functions represent the offset as an s32, which can only represent the
interval [-SZ_2G, SZ_2G - 1], and truncates values beyond this (possibly
altering their sign).

Practically speaking this is unlikely to cause issues outside of
testing, as kernel images are typically far smaller than 2GiB, and
modules are always placed within 2GiB of the kernel image. However, it
may be possible to construct sufficiently large images if a large number
of debug options are enabled, etc.

Fix this by treating the offset as an s64. Since
aarch64_insn_decode_immediate() extracts an unsigned value, we must
sign-extend the immediate after extraction.

Currently aarch64_insn_adrp_get_offset() is the only code which extracts
the ADRP immediate, and we don't have an equivalent function for ADR.

Fixes: 46084bc253e1acdd ("arm64: insn: Add helpers for adrp offsets")
Fixes: 7616fc8bcd7ef497 ("arm64: alternative: Allow immediate branch as alternative instruction")
Cc: Ard Biesheuvel <ardb@kernel.org>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Joey Gouly <joey.gouly@arm.com>
Cc: Marc Zyngier <maz@kernel.org>
Cc: Suzuki K Poulose <suzuki.poulose@arm.com>
Cc: Will Deacon <will@kernel.org>

+++ arch/arm64/include/asm/insn.h
@@ -690,8 +690,8 @@ u32 aarch64_insn_gen_dmb(enum aarch64_insn_mb_type type);
 s32 aarch64_get_branch_offset(u32 insn);
 u32 aarch64_set_branch_offset(u32 insn, s32 offset);
 
-s32 aarch64_insn_adrp_get_offset(u32 insn);
-u32 aarch64_insn_adrp_set_offset(u32 insn, s32 offset);
+s64 aarch64_insn_adrp_get_offset(u32 insn);
+u32 aarch64_insn_adrp_set_offset(u32 insn, s64 offset);
 
 bool aarch32_insn_is_wide(u32 insn);
